auto refresh implemented

+++ Server/src/main/java/com/abs/errp/feedback/FeedbackServiceImpl.java
@@ -47,8 +47,8 @@ public class FeedbackServiceImpl implements FeedbackService {
 		feedback.setSender(e);
 		int receiverId = feedback.getReceiver().getEmployeeId();
 		Optional<ErrpUser> receiver = errpUserRepository.findById(receiverId);
-	
-		if (receiver.get().getSupervisor() != null && receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId())
+		System.out.println(receiver.get().getEmployeeId());
+		if (receiver!=null && receiver.get().getEmployeeId()!=getUser().getEmployeeId() && receiver.get().getSupervisor() != null && receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId())
 			return feedbackRepository.save(feedback);
 		else
 			throw new UnAuthorizedAccessException("User is not Authorized");
@@ -85,7 +85,7 @@ public class FeedbackServiceImpl implements FeedbackService {
 
 		int receiverId = feedback.getReceiver().getEmployeeId();
 		Optional<ErrpUser> receiver = errpUserRepository.findById(receiverId);
-		if (receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId()) {
+		if (receiver!=null && receiver.get().getSupervisor() != null && receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId()) {
 			updateFeedback.setTitle(feedback.getTitle());
 			updateFeedback.setDescription(feedback.getDescription());
 			updateFeedback.setUpdatedDate(new Date());
@@ -106,7 +106,7 @@ public class FeedbackServiceImpl implements FeedbackService {
 			int receiverId = feedback.get().getReceiver().getEmployeeId();
 			Optional<ErrpUser> receiver = errpUserRepository.findById(receiverId);
 
-			if (receiver.get().getSupervisor() != null && receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId())
+			if (receiver!=null && receiver.get().getSupervisor() != null && receiver.get().getSupervisor().getEmployeeId() == getUser().getEmployeeId())
 				this.feedbackRepository.deleteById(id);
 			else
 				throw new UnAuthorizedAccessException("User is not Authorized");
